refactor: remove redundant files in core package
