+++ packages/core/src/shared/types.ts
@@ -17,31 +17,19 @@
 export type Nullable<T> = T | null | undefined | void;
 
 /**
- * wrap any
- */
-export type NoNeedCheckedType = any;
-
-/**
- * Class type
- */
-export interface Class<T> {
-    new (...param: any): T;
-}
-
-/**
  * Key value object
  *
- * @deprecated
+ * @ignore
+ * @deprecated As it has
  */
 export interface IKeyValue {
     [key: string]: any;
 }
 
 /**
- * Custom type of key
+ * @ignore
+ * @deprecated, use {@link Record} instead.
  */
 export interface IKeyType<T> {
     [key: string]: T;
 }
-
-export type AsyncFunction<T = void, R = void> = (value: T) => Promise<R>;
