+++ packages/core/src/types/interfaces/i-shape-properties.ts
@@ -24,10 +24,6 @@ export interface IShapeProperties {
     shapeBackgroundFill: IColorStyle;
     radius?: number;
     outline?: IOutline;
-    // shadow: IShadow;
-    // link: ILink;
-    // contentAlignment: ContentAlignment;
-    // autoFit: IAutoFit;
 }
 
 export interface IOutline {
