+++ packages/core/src/types/interfaces/index.ts
@@ -24,11 +24,8 @@ export * from './i-placeholder';
 export * from './i-range';
 export * from './i-row-data';
 export * from './i-selection-data';
-export * from './i-service-data';
 export * from './i-shape-properties';
 export * from './i-slide-data';
 export * from './i-style-data';
 export * from './i-univer-data';
-export * from './i-workbook-data';
-export * from './i-worksheet-data';
 export * from './i-worksheet-order-data';
